# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Control Plane API
  version: 0.2.0
  summary: Policy, approval, execution, and audit APIs for governed agent actions
  description: |
    v0.2 extends the base control-plane lifecycle with richer, domain-specific schemas for:
      - Git and pull request actions
      - Deploy and cloud actions
      - Email and calendar actions
      - Payment and procurement actions

    Core lifecycle:
      intent -> policy decision -> approval (optional) -> execution grant -> execution -> trace
servers:
  - url: https://api.withaileron.ai
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/connected-accounts:
    get:
      tags: [Connected Accounts]
      summary: List the current user's connected external accounts
      operationId: listConnectedAccounts
      responses:
        "200":
          description: List of connected accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConnectedAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/connected-accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Connected Accounts]
      summary: Get a connected account
      operationId: getConnectedAccount
      responses:
        "200":
          description: Connected account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Connected Accounts]
      summary: Disconnect an external account
      operationId: deleteConnectedAccount
      responses:
        "204":
          description: Account disconnected
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/connect/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          enum: [gmail, google_calendar, outlook, microsoft_calendar, slack, github_repos]
        description: External service provider to connect
    get:
      tags: [Connected Accounts]
      summary: Initiate OAuth flow to connect an external account
      operationId: connectAccount
      description: Redirects the user to the provider's OAuth consent screen with service-specific scopes.
      security: []
      parameters:
        - name: return_to
          in: query
          required: false
          schema: { type: string }
          description: URL to redirect to after the account is connected. Must be a same-site absolute URL or a relative path. Defaults to /settings/connected-accounts.
      responses:
        "302":
          description: Redirect to provider consent screen
        "400":
          $ref: "#/components/responses/BadRequest"
        "423":
          $ref: "#/components/responses/VaultLocked"
  /v1/connect/{provider}/callback:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          enum: [gmail, google_calendar, outlook, microsoft_calendar, slack, github_repos]
    get:
      tags: [Connected Accounts]
      summary: Handle OAuth callback for account connection
      operationId: connectAccountCallback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: false
          schema: { type: string }
      responses:
        "302":
          description: Redirect to connected accounts page on success
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    VaultLocked:
      description: Vault is locked — unlock with POST /v1/users/me/passphrase/unlock
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            request_id: { type: string }
...
    ConnectedAccount:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        user_id:
          type: string
          readOnly: true
        provider:
          type: string
          enum: [gmail, google_calendar, outlook, microsoft_calendar]
          readOnly: true
        external_user_id:
          type: string
          readOnly: true
          description: Provider-specific user identifier (e.g. email for Gmail, username for GitHub, user ID for Slack)
        scopes:
          type: array
          items:
            type: string
          readOnly: true
          description: OAuth scopes granted to Aileron
        status:
          type: string
          enum: [active, expired, revoked]
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...